pyproject.toml
[project]
name = "server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "ag-ui-protocol>=0.1.8",
    "agno>=1.7.12",
    "fastapi[standard]>=0.116.1",
    "google-genai>=1.31.0",
    "prisma>=0.15.0",
    "psycopg2-binary>=2.9.0",
    "python-dotenv>=1.1.1",
    "sqlalchemy>=2.0.0",
    "supabase>=2.18.1",
    "twilio>=9.7.1",
    "vapi-server-sdk>=1.7.0",
]
